commited to this refactor
