+++ ui/frontend/src/lib/pages/settings/RestartButton.svelte
@@ -1,0 +1,7 @@
+<script>
+  export let restart;
+</script>
+
+<span class="rounded hover:bg-red-700 bg-red-500 px-3 py-1 cursor-pointer" on:click={restart}
+  >Restart</span
+>
